pyproject.toml
[tool.poetry]
name = "diabet-detection"
version = "0.1.0"
description = "First package"
authors = ["vladparh <vladparh@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "2.0.0"
pytorch-lightning = "^2.5.0.post0"
pandas = "^2.2.3"
torchmetrics = "^1.6.0"
scikit-learn = "^1.6.0"
hydra-core = "^1.3.2"
torch = {version = "^2.5.1+cu121", source = "pytorch-gpu"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
mlflow = "^2.19.0"
dvc = "^3.58.0"



[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
